Add files via upload
The dataset is in a wide format, with years as columns. For visualization, it’s better to convert it into a long format (one row per year per country). You can use Python (Pandas) or a tool like Excel for this. the Python Code for Preprocessing the other csv file => result